+++ src/main/java/io/zerogone/model/dto/CategoryDto.java
@@ -1,0 +1,13 @@
+package io.zerogone.model.dto;
+
+public class CategoryDto extends DataTransferObject {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
