--- src/main/java/io/zerogone/model/dto/CreateDto.java
@@ -1,5 +1,0 @@
-package io.zerogone.model.dto;
-
-public interface CreateDto<T> {
-    T convertToEntity();
-}
