+++ src/main/java/io/zerogone/model/dto/PostDto.java
@@ -1,0 +1,74 @@
+package io.zerogone.model.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public class PostDto extends DataTransferObject {
+    private String title;
+    private String contents;
+    private UserDto writer;
+    private BlogDto blog;
+    private List<String> categories;
+    @JsonFormat(pattern = "yyyy-MM-dd kk:mm:ss")
+    private LocalDateTime createDateTime;
+    @JsonFormat(pattern = "yyyy-MM-dd kk:mm:ss")
+    private LocalDateTime updateDateTime;
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getContents() {
+        return contents;
+    }
+
+    public void setContents(String contents) {
+        this.contents = contents;
+    }
+
+    public UserDto getWriter() {
+        return writer;
+    }
+
+    public void setWriter(UserDto writer) {
+        this.writer = writer;
+    }
+
+    public BlogDto getBlog() {
+        return blog;
+    }
+
+    public void setBlog(BlogDto blog) {
+        this.blog = blog;
+    }
+
+    public List<String> getCategories() {
+        return categories;
+    }
+
+    public void setCategories(List<String> categories) {
+        this.categories = categories;
+    }
+
+    public LocalDateTime getCreateDateTime() {
+        return createDateTime;
+    }
+
+    public void setCreateDateTime(LocalDateTime createDateTime) {
+        this.createDateTime = createDateTime;
+    }
+
+    public LocalDateTime getUpdateDateTime() {
+        return updateDateTime;
+    }
+
+    public void setUpdateDateTime(LocalDateTime updateDateTime) {
+        this.updateDateTime = updateDateTime;
+    }
+}
